change files hierarchy
