--- plugins/auth-api/core/index.ts
@@ -1,5 +1,0 @@
-export { createUserInfo } from "./create-user-info/create-user-info";
-export { createTokenPayloadInterface } from "./create-token/create-token-payload-interface";
-export { createAuthConstants } from "./create-constants/create-constants";
-export { createTokenService } from "./create-token/create-token-service";
-export { createTokenServiceTests } from "./create-token/create-token-service-test";
