+++ plugins/auth-api/src/core/create-token-service-test.ts
@@ -1,25 +1,22 @@
 import { print } from "recast";
+import { EnumAuthProviderType } from "../types";
 import { DsgContext, Module } from "@amplication/code-gen-types";
-import { removeTSIgnoreComments } from "../../util/ast";
 import { readFile } from "@amplication/code-gen-utils";
-import { EnumAuthProviderType } from "../../types";
 
-export async function createTokenService(
+export async function createTokenServiceTests(
   dsgContext: DsgContext
 ): Promise<Module> {
   const { serverDirectories, resourceInfo } = dsgContext;
+  const authTestsDir = `${serverDirectories.srcDirectory}/tests/auth`;
   const authProvider: EnumAuthProviderType =
     resourceInfo?.settings.authProvider || EnumAuthProviderType.Jwt;
-  const authDir = `${serverDirectories.srcDirectory}/auth`;
   const name =
     authProvider === EnumAuthProviderType.Http ? "Basic" : authProvider;
   const templatePath = require.resolve(
-    `./${name.toLowerCase()}Token.service.template.ts`
+    `../../templates/create-token/${name.toLowerCase()}/${name.toLowerCase()}Token.service.spec.template.ts`
   );
   const file = await readFile(templatePath);
-  const filePath = `${authDir}/base/token.service.base.ts`;
-
-  removeTSIgnoreComments(file);
+  const filePath = `${authTestsDir}/token.service.spec.ts`;
 
   return { code: print(file).code, path: filePath };
 }
