+++ plugins/auth-api/templates/create-constants.template.ts
