+++ plugins/auth-api/templates/create-token/basic/basicToken.service.template.ts
