+++ plugins/auth-api/templates/create-token/constants.ts
