+++ plugins/auth-api/templates/create-token/jwt/jwtToken.service.spec.template.ts
