+++ plugins/auth-api/templates/create-token/jwt/jwtToken.service.template.ts
