+++ plugins/auth-api/templates/create-token/token-payload-interface.template.ts
