+++ plugins/auth-api/templates/default-auth-guard.template.ts
