+++ plugins/auth-api/templates/module.base.template.ts
