+++ plugins/auth-api/templates/user-info.template.ts
